user must be signed in before edits

+++ service_app.py
@@ -1,27 +1,26 @@
-#FB pw serviceApp2019
-from flask import Flask, render_template, request, redirect,jsonify, url_for, flash
-app = Flask(__name__)
+import random
+import string
+import httplib2
+import json
+import requests
 
+from flask import Flask,render_template,request,redirect,jsonify,url_for,flash
+# converts response into object to send to client
+from flask import make_response 
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from my_databasesetup import Base, Service, TaskItem, User
-
 from flask import session as login_session
-import random
-import string
-
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
-import httplib2
-import json
-from flask import make_response # converts response into object to send to client
-import requests
+
+app = Flask(__name__) 
 
 CLIENT_ID = json.loads(
 	open('client_secrets.json', 'r').read())['web']['client_id']
 
-
-engine = create_engine('sqlite:///servicemenuwithusers1.db', connect_args={'check_same_thread': False}, echo=True)
+engine = create_engine(
+	'sqlite:///servicemenuwithusers1.db', connect_args={'check_same_thread': False}, echo=True)
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -30,34 +29,37 @@ session = DBSession()
 # Create anti-forgery state token
 @app.route('/login')
 def showLogin():
-		state = ''.join(random.choice(string.ascii_uppercase + string.digits)
-										for x in xrange(32))
-		login_session['state'] = state
-		return render_template('login.html', STATE = state)
+	state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+					for x in xrange(32))
+	login_session['state'] = state
+	return render_template('login.html', STATE=state)
 
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
-	if request.args.get('state') != login_session['state']: #checks to make sure it is user and not malicious script
+	# checks to make sure it is user and not malicious script
+	if request.args.get('state') != login_session['state']:  
 		response = make_response(json.dumps('Invalid state parameter'), 401)
 		response.headers['Content-Type'] = 'application/json'
 		return response
 	code = request.data
 	try:
-	# upgrade the authorization code into a credentials object
+		# upgrade the authorization code into a credentials object
 		oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
-		oauth_flow.redirect_uri = 'postmessage'	#this is onetime code flow my server is sending off
-		credentials = oauth_flow.step2_exchange(code) #exchanges authorization code for a credentials object
+		# onetime code flow server is sending off
+		oauth_flow.redirect_uri = 'postmessage'
+		# exchanges authorization code for a credentials object	 
+		credentials = oauth_flow.step2_exchange(code) 
 	except FlowExchangeError:
 		response = make_response(
 		json.dumps('Failed to upgrade the authorization code.'), 401)
 		response.headers['Content-Type'] = 'application/json'
 		return response
 
-	  # Check that the access token is valid.
+	# Check that the access token is valid.
 	access_token = credentials.access_token
 	url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
-		   % access_token)
+		         % access_token)
 	h = httplib2.Http()
 	result = json.loads(h.request(url, 'GET')[1])
 	# If there was an error in the access token info, abort.
@@ -86,14 +88,14 @@ def gconnect():
 	stored_gplus_id = login_session.get('gplus_id')
 	if stored_access_token is not None and gplus_id == stored_gplus_id:
 		response = make_response(json.dumps('Current user is already connected.'),
-								 200)
+								            200)
 		response.headers['Content-Type'] = 'application/json'
 		return response
 		
-	 # Store the access token in the session for later use.
+	# Store the access token in the session for later use.
 	login_session['access_token'] = credentials.access_token
 	login_session['gplus_id'] = gplus_id
-	 # Get user info
+	# Get user info
 	userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
 	params = {'access_token': credentials.access_token, 'alt': 'json'}
 	answer = requests.get(userinfo_url, params=params)
@@ -104,7 +106,7 @@ def gconnect():
 	login_session['picture'] = data['picture']
 	login_session['email'] = data['email']
 
-	#see if user exists, if not create new
+	# see if user exists, if not create new
 	user_id = getUserID(login_session['email'])
 	if not user_id:
 		user_id = createUser(login_session)
@@ -114,9 +116,6 @@ def gconnect():
 	output += '<h1>Welcome, '
 	output += login_session['username']
 	output += '!</h1>'
-	output += '<img src="'
-	output += login_session['picture']
-	output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
 	flash("you are now logged in as %s" % login_session['username'])
 	print "done!"
 	return output
@@ -176,7 +175,7 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
 
-#JSON APIs to view Services Information
+# JSON APIs to view Services Information
 
 @app.route('/service/<int:service_id>/task/JSON')
 def serviceTaskJSON(service_id):
@@ -189,7 +188,7 @@ def servicesJSON():
 		services = session.query(Service).all()
 		return jsonify(services= [service.serialize for service in services])
 
-#Show all services, different views per whether user is logged in or not
+# Show all services, different views per whether user is logged in or not
 @app.route('/')
 @app.route('/service/')
 def showServices():
@@ -201,10 +200,10 @@ def showServices():
 		
 		return render_template('show.html', services = services)
 
-#Create a new service
+# Create a new service
 @app.route('/service/new/', methods=['GET','POST'])
 def newService():
-	if request.method == 'POST':
+	if request.method == 'POST' and 'username' in login_session:
 			newService = Service(name = request.form['name'])
 			session.add(newService)
 			#flash('New service %s Successfully Created' % newService.name)
@@ -213,11 +212,11 @@ def newService():
 	else:
 			return render_template('newService.html')
 
-#Edit a service
+# Edit a service
 @app.route('/service/<int:service_id>/edit/', methods = ['GET', 'POST'])
 def editService(service_id):
 	editedService = session.query(Service).filter_by(id = service_id).one()
-	if request.method == 'POST':
+	if request.method == 'POST' and 'username' in login_session:
 			if request.form['name']:
 				editedService.name = request.form['name']
 				#flash('service Successfully Edited %s' % editedService.name)
@@ -226,13 +225,13 @@ def editService(service_id):
 		return render_template('services_edit.html', service = editedService)
 
 
-#Delete a service - not used in application currently but may add later
+# Delete a service - not used in application currently but may add later
 @app.route('/service/<int:service_id>/delete/', methods = ['GET','POST'])
 def deleteService(service_id):
 	serviceToDelete = session.query(Service).filter_by(id = service_id).one()
-	if request.method == 'POST':
+	if request.method == 'POST' and 'username' in login_session:
 		session.delete(serviceToDelete)
-		#flash('%s Successfully Deleted' % serviceToDelete.name)
+		# flash('%s Successfully Deleted' % serviceToDelete.name)
 		session.commit()
 		return redirect(url_for('showServices', service_id = service_id))
 	else:
@@ -248,12 +247,14 @@ def showTask(service_id):
 		 
 
 
-#Create a new Task item
+# Create a new Task item
 @app.route('/service/<int:service_id>/task/new/',methods=['GET','POST'])
 def newTaskItem(service_id):
 	service = session.query(Service).filter_by(id = service_id).one()
-	if request.method == 'POST':
-			newItem = TaskItem(name = request.form['name'], description = request.form['description'], price = request.form['price'], service_id = service_id)
+	if request.method == 'POST' and 'username' in login_session:
+			newItem = TaskItem(
+				name = request.form['name'], description = request.form['description'], 
+				price = request.form['price'], service_id = service_id)
 			session.add(newItem)
 			session.commit()
 			#flash('New Task %s Item Successfully Created' % (newItem.name))
@@ -261,13 +262,13 @@ def newTaskItem(service_id):
 	else:
 			return render_template('newTaskItem.html', service_id = service_id)
 
-# #Edit a Task item
+# Edit a Task item
 @app.route('/service/<int:service_id>/task/<int:task_id>/edit', methods=['GET','POST'])
 def editTaskItem(service_id, task_id):
 
 		editedItem = session.query(TaskItem).filter_by(id = task_id).one()
 		service = session.query(Service).filter_by(id = service_id).one()
-		if request.method == 'POST':
+		if request.method == 'POST' and 'username' in login_session:
 				if request.form['name']:
 						editedItem.name = request.form['name']
 				if request.form['description']:
@@ -281,15 +282,16 @@ def editTaskItem(service_id, task_id):
 				flash('Task Item Successfully Edited')
 				return redirect(url_for('showTask', service_id = service_id))
 		else:
-				return render_template('editTaskitem.html', service_id = service_id, Task_id = task_id, item = editedItem)
+				return render_template(
+					'editTaskitem.html', service_id = service_id, Task_id = task_id, item = editedItem)
 
 
-# #Delete a Task item
-@app.route('/service/<int:service_id>/task/<int:task_id>/delete', methods = ['GET','POST'])
+# Delete a Task item
+@app.route('/service/<int:service_id>/task/<int:task_id>/delete',methods=['GET','POST'])
 def deleteTaskItem(service_id,task_id):
 		service = session.query(Service).filter_by(id = service_id).one()
 		itemToDelete = session.query(TaskItem).filter_by(id = task_id).one() 
-		if request.method == 'POST':
+		if request.method == 'POST' and 'username' in login_session:
 				session.delete(itemToDelete)
 				session.commit()
 				flash('Task Item Successfully Deleted')
@@ -298,11 +300,7 @@ def deleteTaskItem(service_id,task_id):
 				return render_template('deleteTaskItem.html', item = itemToDelete)
 
 
-
-
 if __name__ == '__main__':
 	app.secret_key = 'super_secret_key'
 	app.debug = True
 	app.run(host = '0.0.0.0', port = 5000)
-
-
